kyungpook/day2/day2-2/day2-2_src/requirements.txt
flask==3.0.3
pyjwt==2.8.0